Fix launch and mission bugs
