+++ ui/lib/mission/useMissionData.tsx
@@ -1,9 +1,11 @@
 import LaunchPadPayHookABI from 'const/abis/LaunchPadPayHook.json'
-import { DEFAULT_CHAIN_V5 } from 'const/config'
-import { useEffect, useState } from 'react'
+import { DEFAULT_CHAIN_V5, MISSION_TABLE_NAMES } from 'const/config'
+import { useCallback, useContext, useEffect, useState } from 'react'
 import { readContract, getContract } from 'thirdweb'
 import client from '@/lib/thirdweb/client'
 import useJBProjectData from '../juicebox/useJBProjectData'
+import { getChainSlug } from '../thirdweb/chain'
+import ChainContextV5 from '../thirdweb/chain-context-v5'
 
 /*
 1: Stage 1
@@ -22,10 +24,12 @@ export default function useMissionData({
   projectMetadata,
   projectSubgraphData,
 }: any) {
+  const { selectedChain } = useContext(ChainContextV5)
+  const chainSlug = getChainSlug(selectedChain)
   const [fundingGoal, setFundingGoal] = useState(0)
   const [stage, setStage] = useState<MissionStage>()
   const [backers, setBackers] = useState<any[]>([])
-  const [deadline, setDeadline] = useState<number>(0)
+  const [deadline, setDeadline] = useState<number | undefined>(undefined)
   const [poolDeployerAddress, setPoolDeployerAddress] = useState<string>()
 
   const jbProjectData = useJBProjectData({
@@ -39,21 +43,15 @@ export default function useMissionData({
 
   useEffect(() => {
     async function getFundingData() {
-      const tableName = await readContract({
-        contract: missionTableContract,
-        method: 'getTableName' as string,
-        params: [],
-      })
-
-      const statement = `SELECT * FROM ${tableName} WHERE id = ${mission.id}`
+      const statement = `SELECT * FROM ${MISSION_TABLE_NAMES[chainSlug]} WHERE id = ${mission.id}`
       const res = await fetch(`/api/tableland/query?statement=${statement}`)
       const rows = await res.json()
       setFundingGoal(rows[0]?.fundingGoal)
     }
-    if (missionTableContract && mission?.id !== undefined) {
+    if (mission?.id !== undefined) {
       getFundingData()
     }
-  }, [missionTableContract, mission?.id])
+  }, [mission?.id])
 
   useEffect(() => {
     async function getStage() {
@@ -103,38 +101,41 @@ export default function useMissionData({
         chain: DEFAULT_CHAIN_V5,
         abi: LaunchPadPayHookABI.abi as any,
       })
-      const deadline: any = await readContract({
-        contract: payHookContract,
-        method: 'deadline' as string,
-        params: [],
-      })
-      setDeadline(+deadline.toString() * 1000) // Convert to milliseconds
+      if (payHookContract) {
+        const deadline: any = await readContract({
+          contract: payHookContract,
+          method: 'deadline' as string,
+          params: [],
+        })
+        setDeadline(+deadline.toString() * 1000) // Convert to milliseconds
+      }
     }
     if (missionCreatorContract && mission?.id !== undefined) {
       getDeadline()
     }
   }, [missionCreatorContract, mission?.id])
 
-  //Backers
+  // Backers
+  const refreshBackers = useCallback(async () => {
+    if (mission?.projectId === undefined) return
+
+    const res = await fetch(
+      `/api/mission/backers?projectId=${mission?.projectId}`
+    )
+    const data = await res.json()
+    setBackers(data.backers)
+  }, [mission?.projectId])
+
   useEffect(() => {
-    async function getBackers() {
-      const res = await fetch(
-        `/api/mission/backers?projectId=${mission?.projectId}`
-      )
-      const data = await res.json()
-      setBackers(data.backers)
-    }
-    if (mission?.projectId !== undefined) {
-      getBackers()
-    }
+    refreshBackers()
 
     //Update backers every minute
     const interval = setInterval(() => {
-      getBackers()
+      refreshBackers()
     }, 60000)
 
     return () => clearInterval(interval)
-  }, [mission?.projectId])
+  }, [refreshBackers])
 
   return {
     ...jbProjectData,
@@ -143,5 +144,6 @@ export default function useMissionData({
     backers,
     deadline,
     poolDeployerAddress,
+    refreshBackers,
   }
 }
