+++ ui/pages/mission/[tokenId].tsx
@@ -6,7 +6,8 @@ import JBV4ControllerABI from 'const/abis/JBV4Controller.json'
 import JBV4DirectoryABI from 'const/abis/JBV4Directory.json'
 import JBV4TokenABI from 'const/abis/JBV4Token.json'
 import JBV4TokensABI from 'const/abis/JBV4Tokens.json'
-import MissionCreatorABI from 'const/abis/MissionCreator.json'
+import LaunchPadPayHookABI from 'const/abis/LaunchPadPayHook.json'
+import MissionCreatorABI from 'const/abis/MissionCreatorSep.json'
 import MissionTableABI from 'const/abis/MissionTable.json'
 import PoolDeployerABI from 'const/abis/PoolDeployer.json'
 import TeamABI from 'const/abis/Team.json'
@@ -53,7 +54,7 @@ import queryTable from '@/lib/tableland/queryTable'
 import { useTeamData } from '@/lib/team/useTeamData'
 import { getChainSlug } from '@/lib/thirdweb/chain'
 import ChainContextV5 from '@/lib/thirdweb/chain-context-v5'
-import { serverClient } from '@/lib/thirdweb/client'
+import client, { serverClient } from '@/lib/thirdweb/client'
 import { useChainDefault } from '@/lib/thirdweb/hooks/useChainDefault'
 import useContract from '@/lib/thirdweb/hooks/useContract'
 import useRead from '@/lib/thirdweb/hooks/useRead'
@@ -73,6 +74,7 @@ import { Mission } from '@/components/mission/MissionCard'
 import MissionFundingProgressBar from '@/components/mission/MissionFundingProgressBar'
 import MissionInfo from '@/components/mission/MissionInfo'
 import MissionPayRedeem from '@/components/mission/MissionPayRedeem'
+import MissionProfileHeader from '@/components/mission/MissionProfileHeader'
 import MissionStat from '@/components/mission/MissionStat'
 import { PrivyWeb3Button } from '@/components/privy/PrivyWeb3Button'
 import TeamMembers from '@/components/subscription/TeamMembers'
@@ -80,9 +82,15 @@ import TeamMembers from '@/components/subscription/TeamMembers'
 type ProjectProfileProps = {
   tokenId: string
   mission: Mission
+  stage: number
+  deadline: number | undefined
 }
 
-export default function MissionProfile({ mission }: ProjectProfileProps) {
+export default function MissionProfile({
+  mission,
+  stage,
+  deadline,
+}: ProjectProfileProps) {
   const [isLoading, setIsLoading] = useState(false)
   const account = useActiveAccount()
 
@@ -144,7 +152,7 @@ export default function MissionProfile({ mission }: ProjectProfileProps) {
 
   const missionCreatorContract = useContract({
     address: MISSION_CREATOR_ADDRESSES[chainSlug],
-    abi: MissionCreatorABI.abi as any,
+    abi: MissionCreatorABI as any,
     chain: selectedChain,
   })
 
@@ -154,9 +162,8 @@ export default function MissionProfile({ mission }: ProjectProfileProps) {
     subgraphData,
     fundingGoal,
     primaryTerminalAddress,
-    stage,
     backers,
-    deadline,
+    refreshBackers,
     poolDeployerAddress,
   } = useMissionData({
     mission,
@@ -196,38 +203,11 @@ export default function MissionProfile({ mission }: ProjectProfileProps) {
     params: [account?.address],
   })
 
-  const { data: ethPrice } = useETHPrice(1, 'ETH_TO_USD')
-
-  useEffect(() => {
-    async function getTeamNFT() {
-      if (mission?.teamId === undefined || !teamContract) return
-      const teamNFT = await getNFT({
-        contract: teamContract,
-        tokenId: BigInt(mission.teamId),
-        includeOwner: true,
-      })
-      setTeamNFT({
-        ...teamNFT,
-        metadata: { ...teamNFT.metadata, id: teamNFT.id.toString() },
-      })
-    }
-
-    getTeamNFT()
-  }, [teamContract, mission.teamId])
-
-  useChainDefault()
-
-  const duration = useMemo(() => {
-    console.log('deadline', deadline)
-    return formatTimeUntilDeadline(new Date(deadline))
-  }, [ruleset, deadline])
-
-  const deadlinePassed = Date.now() > deadline
-
   useEffect(() => {
     async function fetchAvailableAmounts() {
       if (
         !jbTerminalContract ||
+        !jbControllerContract ||
         mission?.projectId === undefined ||
         mission?.projectId === null
       )
@@ -359,286 +339,24 @@ export default function MissionProfile({ mission }: ProjectProfileProps) {
     }
   }
 
-  function ProfileHeader() {
-    const totalFunding = useTotalFunding(mission?.projectId)
-    return (
-      <div id="citizenheader-container" className="w-[100vw]">
-        <div className="w-full">
-          <div id="frame-content-container" className="w-full">
-            <div
-              id="frame-content"
-              className="w-full sm:px-[5vw] flex flex-col lg:flex-row items-start xl:px-0 xl:w-[1200px]"
-            >
-              <div
-                id="profile-description-section"
-                className="flex w-full flex-col lg:flex-row items-start lg:items-center"
-              >
-                {mission?.metadata?.logoUri ? (
-                  <div className="pr-0 md:pr-[2vw] pb-[5vw] md:pb-[2vw]">
-                    <div
-                      id="mission-image-container"
-                      className="pl-0 relative w-full h-full md:min-w-[300px] md:min-h-[300px] md:max-w-[300px] md:max-h-[300px]"
-                    >
-                      <IPFSRenderer
-                        src={mission?.metadata?.logoUri}
-                        className="sm:rounded-full rounded-tr-none sm:rounded-tr-full mt-[-3vw] sm:mt-0 w-[100vw] sm:w-full h-full sm:max-w-[350px] sm:max-h-[350px]"
-                        height={576}
-                        width={576}
-                        alt="Mission Image"
-                      />
-                      {teamNFT?.metadata?.image && (
-                        <div
-                          id="team-nft-container"
-                          className="absolute bottom-0 lg:right-0 mb-[-5vw] md:mb-[-2vw] mr-[-5vw] md:mr-[-2vw]"
-                        >
-                          <IPFSRenderer
-                            src={teamNFT?.metadata?.image}
-                            className="top-[2vw] rounded-full ml-[5vw] sm:ml-0"
-                            height={150}
-                            width={150}
-                            alt="Team Image"
-                          />
-                        </div>
-                      )}
-                    </div>
-                  </div>
-                ) : (
-                  <></>
-                )}
-                <div className="flex items-start justify-start w-full sm:w-auto">
-                  <div
-                    id="mission-name"
-                    className="flex px-[5vw] sm:px-0 w-full flex-col justify-center lg:ml-5 max-w-[650px]"
-                  >
-                    <div
-                      id="mission-name-container"
-                      className="mt-5 lg:mt-0 flex flex-col w-full items-start justify-start"
-                    >
-                      {mission ? (
-                        <h1 className="max-w-[450px] text-black opacity-[80%] lg:block font-GoodTimes header dark:text-white text-3xl">
-                          {mission?.metadata?.name}
-                        </h1>
-                      ) : (
-                        <></>
-                      )}
-                    </div>
-                    <div id="profile-container">
-                      {mission?.metadata?.tagline ? (
-                        <p
-                          id="profile-description-container"
-                          className="w-full pr-12 text-gray-300 pb-5"
-                        >
-                          {mission?.metadata?.tagline || ''}
-                        </p>
-                      ) : (
-                        <></>
-                      )}
-                    </div>
+  useEffect(() => {
+    async function getTeamNFT() {
+      if (mission?.teamId === undefined || !teamContract) return
+      const teamNFT = await getNFT({
+        contract: teamContract,
+        tokenId: BigInt(mission.teamId),
+        includeOwner: true,
+      })
+      setTeamNFT({
+        ...teamNFT,
+        metadata: { ...teamNFT.metadata, id: teamNFT.id.toString() },
+      })
+    }
 
-                    {ruleset && teamNFT?.metadata?.name && (
-                      <div className="hidden sm:flex pb-2 flex-col sm:flex-row items-start ">
-                        <p className="opacity-60">
-                          {`Created on ${new Date(
-                            ruleset?.[0]?.start * 1000
-                          ).toLocaleDateString('en-US', {
-                            month: 'long',
-                            day: 'numeric',
-                            year: 'numeric',
-                          })} by: `}
-                        </p>
-                        <Link
-                          href={`/team/${generatePrettyLink(
-                            teamNFT?.metadata?.name
-                          )}`}
-                          className="font-GoodTimes text-white underline sm:pl-2"
-                        >
-                          {teamNFT?.metadata?.name}
-                        </Link>
-                      </div>
-                    )}
+    getTeamNFT()
+  }, [teamContract, mission.teamId])
 
-                    <div className="max-w-[500px] w-full bg-gradient-to-r from-[#3343A5] to-[#18183F] p-4 rounded-xl">
-                      {/* Purple raised amount tag */}
-                      <div className="mb-4 flex flex-col sm:flex-row md:items-center md:justify-between">
-                        <div className="bg-gradient-to-r from-[#51285C] to-[#6D3F79] text-white font-GoodTimes py-2 px-6 rounded-full inline-flex items-start w-fit flex flex-col">
-                          <div className="flex items-center">
-                            <Image
-                              src="/assets/icon-raised-tokens.svg"
-                              alt="Raised"
-                              width={24}
-                              height={24}
-                              className="mr-2"
-                            />
-                            <span className="mr-2">
-                              {truncateTokenValue(
-                                Number(totalFunding || 0) / 1e18,
-                                'ETH'
-                              )}
-                            </span>
-                            <span className="text-sm md:text-base">
-                              ETH RAISED
-                            </span>
-                          </div>
-                          <p className="font-[Lato] text-sm opacity-60">{`($${Math.round(
-                            (Number(totalFunding || 0) / 1e18 || 0) * ethPrice
-                          ).toLocaleString()} USD)`}</p>
-                        </div>
-
-                        {/* Contributors section - visible on md screens and above */}
-                        <div className="hidden sm:flex items-center ml-2 md:mt-0">
-                          <Image
-                            src="/assets/icon-backers.svg"
-                            alt="Backers"
-                            width={24}
-                            height={24}
-                          />
-                          <div className="mx-2">
-                            <p className="sm:hidden text-gray-400 text-sm">
-                              BACKERS
-                            </p>
-                            <p className="text-white font-GoodTimes">
-                              {backers?.length || 0}
-                            </p>
-                          </div>
-                        </div>
-                      </div>
-
-                      <div className="w-full">
-                        <MissionFundingProgressBar
-                          fundingGoal={fundingGoal}
-                          volume={Number(totalFunding || 0) / 1e18}
-                        />
-                      </div>
-
-                      <div className="flex flex-col sm:flex-row flex-wrap gap-4 justify-between sm:justify-start">
-                        <div className="flex items-center">
-                          <Image
-                            src="/assets/launchpad/target.svg"
-                            alt="Goal"
-                            width={24}
-                            height={24}
-                          />
-                          <div className="ml-2">
-                            <div className="flex items-center gap-1">
-                              <p className="text-gray-400 text-sm">GOAL</p>
-                              <Tooltip
-                                text={`~ $${Math.round(
-                                  (fundingGoal / 1e18) * ethPrice
-                                ).toLocaleString()} USD`}
-                                buttonClassName="scale-75"
-                              >
-                                ?
-                              </Tooltip>
-                            </div>
-                            <p className="text-white font-GoodTimes">
-                              {+(fundingGoal / 1e18).toFixed(3)} ETH
-                            </p>
-                          </div>
-                        </div>
-
-                        <div className="flex items-center">
-                          <Image
-                            src="/assets/launchpad/clock.svg"
-                            alt="Deadline"
-                            width={24}
-                            height={24}
-                          />
-                          <div className="ml-2">
-                            <p className="text-gray-400 text-sm">DEADLINE</p>
-                            <p className="text-white font-GoodTimes">
-                              {duration}
-                            </p>
-                          </div>
-                        </div>
-
-                        {/* Contributors section - visible only on smaller screens */}
-                        <div className="flex sm:hidden items-center">
-                          <Image
-                            src="/assets/icon-backers.svg"
-                            alt="Backers"
-                            width={24}
-                            height={24}
-                          />
-                          <div className="ml-2">
-                            <p className="text-gray-400 text-sm">BACKERS</p>
-                            <p className="text-white font-GoodTimes">
-                              {backers?.length || 0}
-                            </p>
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-
-                    {ruleset && teamNFT?.metadata?.name && (
-                      <div className="flex sm:hidden pt-2 flex-col sm:flex-row items-start ">
-                        <p className="opacity-60">
-                          {`Created on ${new Date(
-                            ruleset?.[0]?.start * 1000
-                          ).toLocaleDateString('en-US', {
-                            month: 'long',
-                            day: 'numeric',
-                            year: 'numeric',
-                          })} by: `}
-                        </p>
-                        <Link
-                          href={`/team/${generatePrettyLink(
-                            teamNFT?.metadata?.name
-                          )}`}
-                          className="font-GoodTimes text-white underline"
-                        >
-                          {teamNFT?.metadata?.name}
-                        </Link>
-                      </div>
-                    )}
-                    {/* Send payouts and tokens Buttons - only shown to managers */}
-                    {account && deadlinePassed && isManager && (
-                      <div className="flex flex-col gap-4 mt-8 md:-mt-8 w-full sm:w-auto sm:absolute sm:right-2 sm:top-[250px]">
-                        <PrivyWeb3Button
-                          requiredChain={DEFAULT_CHAIN_V5}
-                          className="gradient-2 rounded-full w-full noPadding leading-none flex-1 sm:w-[250px]"
-                          label={
-                            <span className="whitespace-nowrap">
-                              Send Tokens
-                            </span>
-                          }
-                          action={sendReservedTokens}
-                          isDisabled={!availableTokens}
-                        />
-                        <PrivyWeb3Button
-                          requiredChain={DEFAULT_CHAIN_V5}
-                          className="gradient-2 rounded-full noPadding w-full leading-none flex-1 sm:w-[250px]"
-                          label={
-                            <span className="whitespace-nowrap">
-                              Send Payouts
-                            </span>
-                          }
-                          action={sendPayouts}
-                          isDisabled={!availablePayouts}
-                        />
-                        {stage === 2 && (
-                          <PrivyWeb3Button
-                            requiredChain={DEFAULT_CHAIN_V5}
-                            className="gradient-2 rounded-full noPadding w-full leading-none flex-1 sm:w-[250px]"
-                            label={
-                              <span className="whitespace-nowrap">
-                                Deploy Liquidity
-                              </span>
-                            }
-                            action={deployLiquidityPool}
-                            isDisabled={!poolDeployerAddress}
-                          />
-                        )}
-                      </div>
-                    )}
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    )
-  }
+  useChainDefault()
 
   return (
     <JuiceProviders
@@ -654,7 +372,24 @@ export default function MissionProfile({ mission }: ProjectProfileProps) {
         <ContentLayout
           header={''}
           headerSize="max(20px, 3vw)"
-          description={<ProfileHeader />}
+          description={
+            <MissionProfileHeader
+              mission={mission}
+              teamNFT={teamNFT}
+              ruleset={ruleset}
+              fundingGoal={fundingGoal}
+              backers={backers}
+              deadline={deadline}
+              stage={stage}
+              poolDeployerAddress={poolDeployerAddress}
+              isManager={isManager}
+              availableTokens={availableTokens}
+              availablePayouts={availablePayouts}
+              sendReservedTokens={sendReservedTokens}
+              sendPayouts={sendPayouts}
+              deployLiquidityPool={deployLiquidityPool}
+            />
+          }
           mainPadding
           mode="compact"
           popOverEffect={false}
@@ -691,6 +426,7 @@ export default function MissionProfile({ mission }: ProjectProfileProps) {
                     primaryTerminalAddress={primaryTerminalAddress}
                     jbControllerContract={jbControllerContract}
                     jbTokensContract={jbTokensContract}
+                    refreshBackers={refreshBackers}
                   />
                 ) : (
                   <div className="p-4 text-center">
@@ -763,60 +499,103 @@ export default function MissionProfile({ mission }: ProjectProfileProps) {
 }
 
 export const getServerSideProps: GetServerSideProps = async ({ params }) => {
-  const tokenId: any = params?.tokenId
+  try {
+    const tokenId: any = params?.tokenId
 
-  const chain = DEFAULT_CHAIN_V5
-  const chainSlug = getChainSlug(chain)
+    const chain = DEFAULT_CHAIN_V5
+    const chainSlug = getChainSlug(chain)
 
-  if (tokenId === undefined) {
+    if (tokenId === undefined || isNaN(Number(tokenId))) {
+      return {
+        notFound: true,
+      }
+    }
+
+    const missionTableName = MISSION_TABLE_NAMES[chainSlug]
+
+    const statement = `SELECT * FROM ${missionTableName} WHERE id = ${tokenId}`
+
+    const missionRows = await queryTable(chain, statement)
+    const missionRow = missionRows?.[0]
+
+    if (!missionRow || blockedMissions.includes(Number(tokenId))) {
+      return {
+        notFound: true,
+      }
+    }
+
+    const jbV4ControllerContract = getContract({
+      client: serverClient,
+      address: JBV4_CONTROLLER_ADDRESSES['sepolia'],
+      abi: JBV4ControllerABI as any,
+      chain: DEFAULT_CHAIN_V5,
+    })
+
+    const metadataURI = await readContract({
+      contract: jbV4ControllerContract,
+      method: 'uriOf' as string,
+      params: [missionRow.projectId],
+    })
+
+    const ipfsHash = metadataURI.startsWith('ipfs://')
+      ? metadataURI.replace('ipfs://', '')
+      : metadataURI
+
+    const metadataRes = await fetch(`${IPFS_GATEWAY}${ipfsHash}`)
+    const metadata = await metadataRes.json()
+
+    const mission = {
+      id: missionRow.id,
+      teamId: missionRow.teamId,
+      projectId: missionRow.projectId,
+      metadata: metadata,
+    }
+
+    const missionCreatorContract = getContract({
+      client: serverClient,
+      address: MISSION_CREATOR_ADDRESSES['sepolia'],
+      abi: MissionCreatorABI as any,
+      chain: DEFAULT_CHAIN_V5,
+    })
+
+    const stage = await readContract({
+      contract: missionCreatorContract,
+      method: 'stage' as string,
+      params: [mission.id],
+    })
+
+    const payHookAddress: any = await readContract({
+      contract: missionCreatorContract,
+      method: 'missionIdToPayHook' as string,
+      params: [mission.id],
+    })
+    const payHookContract = getContract({
+      client,
+      address: payHookAddress,
+      chain: DEFAULT_CHAIN_V5,
+      abi: LaunchPadPayHookABI.abi as any,
+    })
+    let deadline: number | undefined
+    if (payHookContract) {
+      const dl: any = await readContract({
+        contract: payHookContract,
+        method: 'deadline' as string,
+        params: [],
+      })
+      deadline = +dl.toString() * 1000 // Convert to milliseconds
+    }
+
+    return {
+      props: {
+        mission,
+        stage: +stage.toString(),
+        deadline,
+      },
+    }
+  } catch (error) {
+    console.error(error)
     return {
       notFound: true,
     }
-  }
-
-  const missionTableName = MISSION_TABLE_NAMES[chainSlug]
-
-  const statement = `SELECT * FROM ${missionTableName} WHERE id = ${tokenId}`
-
-  const missionRows = await queryTable(chain, statement)
-  const missionRow = missionRows?.[0]
-
-  if (!missionRow || blockedMissions.includes(Number(tokenId))) {
-    return {
-      notFound: true,
-    }
-  }
-
-  const jbV4ControllerContract = getContract({
-    client: serverClient,
-    address: JBV4_CONTROLLER_ADDRESSES[chainSlug],
-    abi: JBV4ControllerABI as any,
-    chain: chain,
-  })
-
-  const metadataURI = await readContract({
-    contract: jbV4ControllerContract,
-    method: 'uriOf' as string,
-    params: [missionRow.projectId],
-  })
-
-  const ipfsHash = metadataURI.startsWith('ipfs://')
-    ? metadataURI.replace('ipfs://', '')
-    : metadataURI
-
-  const metadataRes = await fetch(`${IPFS_GATEWAY}${ipfsHash}`)
-  const metadata = await metadataRes.json()
-
-  const mission = {
-    id: missionRow.id,
-    teamId: missionRow.teamId,
-    projectId: missionRow.projectId,
-    metadata: metadata,
-  }
-
-  return {
-    props: {
-      mission,
-    },
   }
 }
